Fix issue #2 (static app example using Opal::Environment)
Opal::Environment has been deprecated and should not be used. But the
example is also buggy in that it does not require
sprockets/environment. Once you fix that, you get the deprecation
warning to use Opal::Server which does not apply to building static
apps. So, the solution is to use Opal::Builder for this task instead.