Update libraries page with latest upstream content
Prior to this commit, Opal linked to Hyperloop, which is gone, and Hyperstack is its replacement. This is on the prominent "libraries" page, which a lot of people may visit, and find that one of the biggest projects now has a defunct domain. This commit updates the libraries page with the latest content from https://github.com/fazibear/awesome-opal, which also fixes this problem.